# Speed up osmosis_highway_deadend sql20