@frodrigo 2023-12-09
The inner `JOIN nodes` could be moved outer the LEFT JOIN. Probably only a small improvement.

As it use a `Materialize` using temps table with index make sens for ferry and parking.