@Famlam 2023-12-09
> The inner JOIN nodes could be moved outer the LEFT JOIN. Probably only a small improvement.

I'm not sure if I understand what you mean, but this would probably cause the `COUNT(*)=1` to fail in case you have a connection between an end node of a highway and a node somewhere in the middle of a highway (like in a "T")

> As it use a Materialize using temps table with index make sens for ferry and parking.

I'll give it a try in a couple of days.