@Famlam 2023-12-09
> I'll give it a try in a couple of days.

Ok, couple of hours actually :)

With ferry and bicycle_parking in their own `TEMP TABLE`, and indices on their `linestring`
japan_chubu: 58 sec | 31 sec
netherlands_gelderland: 47 sec | 12 sec
usa_iowa: 25 sec | 13 sec
slovenia: 22 sec | 6 sec

No clue why the second (and all subsequent) run is so much faster this time, possibly osm110 was busy the first time. But all of the first runs seem an improvement already.

---
japan_chubu
```
['Index Scan using idx_ways_tags on ways  (cost=0.41..157.21 rows=1 width=242) (actual time=48.135..2894.786 rows=53 loops=1)']
["  Index Cond: (tags ? 'route'::text)"]
['  Rows Removed by Index Recheck: 416646']
["  Filter: ((tags -> 'route'::text) = 'ferry'::text)"]
['  Rows Removed by Filter: 668']
['Planning Time: 2.387 ms']
['Execution Time: 2894.888 ms']

['Bitmap Heap Scan on ways  (cost=916.57..63023.39 rows=255 width=242) (actual time=283.111..1594.082 rows=1640 loops=1)']
["  Recheck Cond: ((tags ? 'amenity'::text) AND (tags <> ''::hstore))"]
['  Rows Removed by Index Recheck: 19762']
["  Filter: ((tags -> 'amenity'::text) = 'bicycle_parking'::text)"]
['  Rows Removed by Filter: 84792']
['  Heap Blocks: exact=46209']
['  ->  Bitmap Index Scan on idx_ways_tags  (cost=0.00..916.50 rows=50945 width=0) (actual time=273.198..273.198 rows=106194 loops=1)']
["        Index Cond: (tags ? 'amenity'::text)"]
['Planning Time: 0.212 ms']
['Execution Time: 1594.798 ms']

['Nested Loop Left Join  (cost=39497013.07..39497082.35 rows=1 width=80) (actual time=39986.298..40222.355 rows=371 loops=1)']
['  Filter: (ferry.* IS NULL)']
['  ->  GroupAggregate  (cost=39497012.94..39497068.31 rows=11 width=80) (actual time=39986.236..40206.376 rows=371 loops=1)']
['        Group Key: nodes.id']
['        Filter: (count(*) = 1)']
['        Rows Removed by Filter: 105044']
['        ->  Sort  (cost=39497012.94..39497018.48 rows=2215 width=58) (actual time=39986.197..40052.745 rows=522766 loops=1)']
['              Sort Key: nodes.id']
['              Sort Method: external merge  Disk: 32664kB']
['              ->  Nested Loop  (cost=0.98..39496889.86 rows=2215 width=58) (actual time=1.237..39281.533 rows=522766 loops=1)']
['                    ->  Nested Loop  (cost=0.41..39490419.21 rows=2216 width=26) (actual time=0.927..26424.495 rows=522824 loops=1)']
['                          ->  Seq Scan on highway_ends way_ends  (cost=0.00..222209.65 rows=210945 width=183) (actual time=0.208..2820.594 rows=206932 loops=1)']
["                                Filter: ((level < 3) OR (highway = 'cycleway'::text))"]
['                                Rows Removed by Filter: 3681914']
['                          ->  Index Scan using idx_highways_linestring on highways  (cost=0.41..186.05 rows=10 width=227) (actual time=0.058..0.113 rows=3 loops=206932)']
['                                Index Cond: (linestring && way_ends.linestring)']
['                                Filter: ((NOT is_construction) AND (way_ends.nid = ANY (nodes)))']
['                                Rows Removed by Filter: 21']
['                    ->  Index Scan using pk_nodes on nodes  (cost=0.56..2.92 rows=1 width=40) (actual time=0.024..0.024 rows=1 loops=522824)']
['                          Index Cond: (id = way_ends.nid)']
["                          Filter: (((NOT (tags ? 'amenity'::text)) OR ((tags -> 'amenity'::text) <> 'bicycle_parking'::text)) AND ((NOT (tags ? 'entrance'::text)) OR ((tags -> 'entrance'::text) = 'no'::text)) AND ((NOT (tags ? 'noexit'::text)) OR ((tags -> 'noexit'::text) = 'no'::text)))"]
['                          Rows Removed by Filter: 0']
['  ->  Index Scan using idx_ferry_linestring on ferry  (cost=0.14..1.26 rows=1 width=152) (actual time=0.042..0.042 rows=0 loops=371)']
['        Index Cond: (linestring && nodes.geom)']
['        Filter: (nodes.id = ANY (nodes))']
['        Rows Removed by Filter: 3']
['Planning Time: 12.983 ms']
['Execution Time: 40224.434 ms']

['Nested Loop Left Join  (cost=0.14..33.37 rows=1 width=48) (actual time=0.496..2.694 rows=136 loops=1)']
['  Filter: (bicycle_parking.* IS NULL)']
['  Rows Removed by Filter: 4']
['  ->  Seq Scan on unconnected_highways  (cost=0.00..19.38 rows=4 width=48) (actual time=0.013..0.073 rows=140 loops=1)']
["        Filter: (highway = 'cycleway'::text)"]
['        Rows Removed by Filter: 231']
['  ->  Index Scan using idx_bicycle_parking_linestring on bicycle_parking  (cost=0.14..3.18 rows=1 width=152) (actual time=0.014..0.014 rows=0 loops=140)']
['        Index Cond: (linestring && unconnected_highways.geom)']
['        Filter: (unconnected_highways.nid = ANY (nodes))']
['Planning Time: 0.290 ms']
['Execution Time: 2.941 ms']

['Seq Scan on unconnected_highways  (cost=0.00..951.88 rows=746 width=48) (actual time=0.008..0.145 rows=231 loops=1)']
["  Filter: (highway <> 'cycleway'::text)"]
['  Rows Removed by Filter: 140']
['Planning Time: 0.018 ms']
['Execution Time: 0.164 ms']
```